+++ index.js
@@ -1,15 +1,26 @@
 import Book from './modules/book.js';
 import div from './modules/div.js';
-import { createRemove, addBookToList, removeContent, removeBook } from './modules/bookControllers.js';
+import addBookToList from './modules/bookControllers.js';
 import emptyValue from './modules/emptyValue.js';
 
-export const book = new Book();
+const book = new Book();
 
 const titleInput = document.getElementsByName('title')[0];
 const authorInput = document.getElementsByName('author')[0];
 const addBtn = document.getElementsByName('add')[0];
 const erroMsg = document.querySelector('.error-msg');
 const booksContainer = document.querySelector('.display-book');
+
+const removeContent = (text) => {
+  const element = document.querySelector(text);
+  element.remove();
+};
+export const removeBook = (btn) => {
+  btn.addEventListener('click', (event) => {
+    book.removeBook(Number(event.target.id));
+    removeContent(`.item${event.target.id}`);
+  });
+};
 
 if (addBtn) {
   addBtn.addEventListener('click', () => {
@@ -82,3 +93,5 @@ const options = {
 const formattedDate = currentDate.toLocaleString('en-US', options);
 
 document.querySelector('.time').innerHTML = formattedDate;
+
+export default book;
